+++ mas-foundation/src/main/java/com/ca/mas/core/context/DeviceIdentifier.java
@@ -7,21 +7,23 @@
  */
 package com.ca.mas.core.context;
 
-import android.content.Context;
-
 import com.ca.mas.core.security.KeyStoreException;
 
 import java.security.NoSuchAlgorithmException;
+
+/**
+ * Generates a set of asymmetric keys in the Android keystore and builds the device identifier off of the public key.
+ * Apps built with the same sharedUserId value in AndroidManifest.xml will reuse the same identifier.
+ */
 
 public class DeviceIdentifier extends UniqueIdentifier {
 
     /**
      * Generates a set of asymmetric keys in the Android keystore and builds the device identifier off of the public key.
      * Apps built with the same sharedUserId value in AndroidManifest.xml will reuse the same identifier.
-     * @param context
      */
-    public DeviceIdentifier(Context context) throws KeyStoreException, NoSuchAlgorithmException {
-        super(context);
+    public DeviceIdentifier() throws KeyStoreException, NoSuchAlgorithmException {
+        super();
     }
 
     @Override